requirements.txt
django_cron == 0.4.6
djangorestframework == 3.3.3
Django == 1.9.5